+++ backend/apps/accounts/templates/hospital_signup.html
@@ -32,7 +32,7 @@ Registrierung - Studierende
 	
 
     <h1 class="font-weight-light match4hc-title"> {% blocktrans %}Sie brauchen in Ihrer Institution aktuell Hilfe?{% endblocktrans %} </h1>
-	<h2 class="font-weight-light match4hc-title">Wir verbinden Helfende und Hilfesuchenden zu einem Match4Healthcare!</h2>
+	<h2 class="font-weight-light match4hc-title">{% blocktrans %}Wir verbinden Helfende und Hilfesuchenden zu einem match4Healthcare!{% endblocktrans %}</h2>
        
             <div class="row" style="margin-top:30px;">
                 <div class="col-md-8 col-sm-10 col-12">
