+++ backend/apps/accounts/templates/student_signup.html
@@ -68,7 +68,7 @@ Registrierung - Studierende
 	
 
     <h1 class="font-weight-light match4hc-title">  {% blocktrans %}Du sitzt zuhause und kannst in der aktuellen Situation helfen?{% endblocktrans %}  </h1>
-	<h2 class="font-weight-light match4hc-title">Wir verbinden Helfende und Hilfesuchenden zu einem Match4Healthcare!</h2>
+	<h2 class="font-weight-light match4hc-title">{% blocktrans %}Wir verbinden Helfende und Hilfesuchenden zu einem match4Healthcare!{% endblocktrans %}</h2>
