+++ backend/manage.py
@@ -5,7 +5,7 @@ import sys
 
 
 def main():
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Match4healthcare.settings.development')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'match4healthcare.settings.development')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
