+++ backend/match4healthcare/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for Match4healthcare project.
+ASGI config for match4healthcare project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Match4healthcare.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'match4healthcare.settings')
 
 application = get_asgi_application()
